feat: add support for sending additional data in push notifications
The `SendNotification` method in the `PushNotificationService` now accepts an additional parameter `data` which allows sending custom data along with the push notification. This will enable the client to handle different types of notifications based on the received data. 

A new class `NotificationType` has been added to define constants for different types of notifications. This will help in maintaining consistency and avoiding hardcoding of notification types.

In the `AddOrder` mutation of the GraphQL API, the push notification sent to waiters now includes the type of notification as part of the data. This will allow the client to differentiate between different types of notifications and take appropriate actions.

In the `FinishOrder` mutation of the GraphQL API, the push notification sent to waiters now includes the type of notification as part of the data. This will allow the client to differentiate between different types of notifications and take appropriate actions.